# services/scenario-manager/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Scenario Manager API",
    "description": "API for managing scenarios, strategies, and investments.",
    "version": "v1"
  },
  "paths": {
...
    "/api/scenarios": {
      "post": {
        "tags": [
          "Scenarios"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScenarioRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScenarioRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScenarioRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Scenarios"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScenarioResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScenarioResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScenarioResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/scenarios/{id}": {
      "get": {
        "tags": [
          "Scenarios"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Scenarios"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScenarioRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScenarioRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScenarioRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Scenarios"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/scenarios/{scenarioId}/strategies": {
      "post": {
        "tags": [
          "Strategies"
        ],
        "parameters": [
          {
            "name": "scenarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStrategyRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStrategyRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStrategyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Strategies"
        ],
        "parameters": [
          {
            "name": "scenarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StrategyResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StrategyResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StrategyResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/scenarios/{scenarioId}/strategies/{id}": {
      "get": {
        "tags": [
          "Strategies"
        ],
        "parameters": [
          {
            "name": "scenarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Strategies"
        ],
        "parameters": [
          {
            "name": "scenarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStrategyRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStrategyRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStrategyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Strategies"
        ],
        "parameters": [
          {
            "name": "scenarioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateScenarioRequest": {
        "required": [
          "name",
          "scenarioType"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "scenarioType": {
            "$ref": "#/components/schemas/ScenarioType"
          }
        },
        "additionalProperties": false
      },
      "CreateStrategyRequest": {
        "required": [
          "name",
          "scenarioId"
        ],
        "type": "object",
        "properties": {
          "scenarioId": {
            "minLength": 1,
            "type": "string"
          },
          "strategyType": {
            "$ref": "#/components/schemas/StrategyType"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "maxLength": 1000,
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "ordering": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Distribution": {
        "required": [
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "mean": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "stdev": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "lower": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "upper": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ScenarioResponse": {
        "required": [
          "id",
          "name",
          "ownerId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "ownerId": {
            "type": "string",
            "nullable": true
          },
          "scenarioType": {
            "$ref": "#/components/schemas/ScenarioType"
          },
          "status": {
            "$ref": "#/components/schemas/ScenarioStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "userBirthYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "spouseBirthYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "userLifeExpectancy": {
            "$ref": "#/components/schemas/Distribution"
          },
          "spouseLifeExpectancy": {
            "$ref": "#/components/schemas/Distribution"
          },
          "financialGoal": {
            "type": "number",
            "format": "double"
          },
          "stateOfResidence": {
            "type": "string",
            "nullable": true
          },
          "inflationAssumption": {
            "$ref": "#/components/schemas/Distribution"
          },
          "annualRetirementContributionLimit": {
            "type": "number",
            "format": "double"
          },
          "rothOptimizerEnabled": {
            "type": "boolean"
          },
          "rothOptimizerStartYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "rothOptimizerEndYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "importSource": {
            "type": "string",
            "nullable": true
          },
          "exportCount": {
            "type": "integer",
            "format": "int32"
          },
          "lastSimulationRun": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ScenarioStatus": {
        "enum": [
          "Draft",
          "Complete"
        ],
        "type": "string"
      },
      "ScenarioType": {
        "enum": [
          "Individual",
          "MarriedCouple"
        ],
        "type": "string"
      },
...
      "StrategyResponse": {
        "required": [
          "id",
          "name",
          "scenarioId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "scenarioId": {
            "minLength": 1,
            "type": "string"
          },
          "strategyType": {
            "$ref": "#/components/schemas/StrategyType"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "maxLength": 1000,
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "ordering": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StrategyType": {
        "enum": [
          "Spending",
          "ExpenseWithdrawal",
          "Rmd",
          "RothConversion"
        ],
        "type": "string"
      },
...
      "UpdateScenarioRequest": {
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 255,
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "scenarioType": {
            "$ref": "#/components/schemas/ScenarioType"
          },
          "userBirthYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "spouseBirthYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "userLifeExpectancy": {
            "$ref": "#/components/schemas/Distribution"
          },
          "spouseLifeExpectancy": {
            "$ref": "#/components/schemas/Distribution"
          },
          "financialGoal": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "stateOfResidence": {
            "maxLength": 100,
            "type": "string",
            "nullable": true
          },
          "inflationAssumption": {
            "$ref": "#/components/schemas/Distribution"
          },
          "annualRetirementContributionLimit": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "rothOptimizerEnabled": {
            "type": "boolean",
            "nullable": true
          },
          "rothOptimizerStartYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "rothOptimizerEndYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}